ethereum-spare-part-management/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ethereum-spare-part-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anvil>=0.0.2",
    "ape-alchemy>=0.8.10",
    "ape-foundry>=0.8.10",
    "ape-solidity>=0.8.5",
    "eth-ape>=0.8.41",
    "fastapi>=0.123.0",
    "requests>=2.32.5",
    "setuptools>=80.9.0",
    "streamlit>=1.51.0",
    "uvicorn>=0.38.0",
    "web3>=7.14.0",
]
